Optionally return a new object on find
By default, FrozenRecord objects are instantiated once and then cached
indefinitely in memory. Each time a finder method is called, the same
cached object is returned. As a result, any state that's set on the
FrozenRecord object (in instance variables, for example) is shared state
and will remain present on that object when it is next returned from a
finder method.

This behavior can be unintuitive for users of ActiveRecord, since each
time an ActiveRecord finder method is called a new object is
instantiated with its own separate state.

This adds a new option that allows users who are storing state on their
objects and prefer the ActiveRecord behavior to receive a new object
each time they call a finder method.